Modify labels to match design
Note that since we unforunately have to use the labels to track what we
select on the cascader, this required some string manipulation to
implement. At the core of it, we append '(1)' counts to labels for the
ui and then remove them when sending a request. Luckily this doesn't
effect the lowest level of the cascader which is what external
fucntions use to modify the cascader